fix(app): fixes poap and eth posession sidebar content post platform combination (#705)
* fix(app): fixes poap and eth posession sidebar content

* fix(app): fixes poap and eth posession sidebar content

* fix(app): fixes poap and eth posession sidebar content

* fix(app): fixes poap and eth posession sidebar content
